upstream: move static and strict_dns clusters to dedicated files
Code movement only. The previous entangelment makes it hard to
see specific cluster implementations vs. common code.
