+++ source/common/upstream/static_cluster.cc
@@ -1,0 +1,67 @@
+#include "common/upstream/static_cluster.h"
+
+namespace Envoy {
+namespace Upstream {
+
+StaticClusterImpl::StaticClusterImpl(
+    const envoy::api::v2::Cluster& cluster, Runtime::Loader& runtime,
+    Server::Configuration::TransportSocketFactoryContext& factory_context,
+    Stats::ScopePtr&& stats_scope, bool added_via_api)
+    : ClusterImplBase(cluster, runtime, factory_context, std::move(stats_scope), added_via_api),
+      priority_state_manager_(
+          new PriorityStateManager(*this, factory_context.localInfo(), nullptr)) {
+  // TODO(dio): Use by-reference when cluster.hosts() is removed.
+  const envoy::api::v2::ClusterLoadAssignment cluster_load_assignment(
+      cluster.has_load_assignment() ? cluster.load_assignment()
+                                    : Config::Utility::translateClusterHosts(cluster.hosts()));
+
+  overprovisioning_factor_ = PROTOBUF_GET_WRAPPED_OR_DEFAULT(
+      cluster_load_assignment.policy(), overprovisioning_factor, kDefaultOverProvisioningFactor);
+
+  for (const auto& locality_lb_endpoint : cluster_load_assignment.endpoints()) {
+    priority_state_manager_->initializePriorityFor(locality_lb_endpoint);
+    for (const auto& lb_endpoint : locality_lb_endpoint.lb_endpoints()) {
+      priority_state_manager_->registerHostForPriority(
+          "", resolveProtoAddress(lb_endpoint.endpoint().address()), locality_lb_endpoint,
+          lb_endpoint);
+    }
+  }
+}
+
+void StaticClusterImpl::startPreInit() {
+  // At this point see if we have a health checker. If so, mark all the hosts unhealthy and
+  // then fire update callbacks to start the health checking process.
+  const auto& health_checker_flag =
+      health_checker_ != nullptr
+          ? absl::optional<Upstream::Host::HealthFlag>(Host::HealthFlag::FAILED_ACTIVE_HC)
+          : absl::nullopt;
+
+  auto& priority_state = priority_state_manager_->priorityState();
+  for (size_t i = 0; i < priority_state.size(); ++i) {
+    if (priority_state[i].first == nullptr) {
+      priority_state[i].first = std::make_unique<HostVector>();
+    }
+    priority_state_manager_->updateClusterPrioritySet(
+        i, std::move(priority_state[i].first), absl::nullopt, absl::nullopt, health_checker_flag,
+        overprovisioning_factor_);
+  }
+  priority_state_manager_.reset();
+
+  onPreInitComplete();
+}
+
+ClusterImplBaseSharedPtr StaticClusterFactory::createClusterImpl(
+    const envoy::api::v2::Cluster& cluster, ClusterFactoryContext& context,
+    Server::Configuration::TransportSocketFactoryContext& socket_factory_context,
+    Stats::ScopePtr&& stats_scope) {
+  return std::make_unique<StaticClusterImpl>(cluster, context.runtime(), socket_factory_context,
+                                             std::move(stats_scope), context.addedViaApi());
+}
+
+/**
+ * Static registration for the static cluster factory. @see RegisterFactory.
+ */
+REGISTER_FACTORY(StaticClusterFactory, ClusterFactory);
+
+} // namespace Upstream
+} // namespace Envoy
